Pass the login state parameter to the server-side login context
Previously the `state` parameter for the `/auth/authorize` endpoint was
only stored in memory in the front-end, and never passed to the server.
This updates the code to pass it to the server when creating the login
flow.

This allows the server to remember the state, which can be used to
support auth providers that temporarily redirect the user to an external
server to perform authentication, such as OAuth 2 or OpenID Connect.

Just to be explicit and help avoid confusion, the existing `state`
parameter is sent to us by other clients that want to use Home Assistant
as an OAuth 2 Authentication Server. This commit is changing the code in
order to support Home Assistant acting as an OAuth 2 client, and relying
on an external Authentication Server to perform authentication.