Add support for an error parameter to the authorize endpoint
This parameter causes the `/auth/authorize` endpoint to display the
specified error message to the user until their next login attempt.

This functionality allows auth providers that use an external
redirection step to redirect the user back to the login page with an
error message if their external login attempt fails.